add decryptor.py
Thanks for your work! However, as a security company, you should not invite
users to download an unsigned executable from the internet and run it as
Administrator. This is even more dangerous in an incident response setting.
For this reason, we're providing the source code to the users of this tool,
so they know what they're running on their boxes.

Additionally, this application is vulnerable to race condition that can
cause an elevation of privilege. The application creates the file `pd.exe`
in the temporary folder, then runs it. If the application runs as `SYSTEM`,
it creates `pd.exe` in `c:\windows\temp`, which is writeable by any user.
Consequently, a malicious local user (the ransomware author?) could end up
leveraging this tool to gain privileges or persistance on the target host.

There's also a command injection bug at line 58, but it doesn't appear to
cause security issues.

Finally, we recommend using GitHub releases to host executables. Commits
should be reserved to code.

Love, JBZ